Create RunTime_Interpreter
Created the Interpreter Method which looks for the tags of the symbols in the tokens list and perform operations. 
Implemented Stack [a list which appends and pop to implement LIFO], a symbol table {"" : ""} which holds the key value pairs for x = 5 where x is key and 5 is value.
Implemented Push Operation in the stack and handled the case when the tag is Assign and Operator.